requirements.txt
# ─── Core ──────────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ─── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# ─── ML & Data Science ─────────────────────────────────────────────────────────
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
statsmodels==0.14.0
pmdarima==2.0.4        # auto_arima for SARIMA parameter search
mlflow==2.9.2          # experiment tracking & model registry
joblib==1.3.2          # model serialization
scipy==1.11.4

# ─── API ───────────────────────────────────────────────────────────────────────
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0   # JWT
passlib[bcrypt]==1.7.4             # password hashing
python-multipart==0.0.6
apscheduler==3.10.4                # background job scheduler
slowapi==0.1.9                     # rate limiting
httpx==0.25.2                      # async HTTP client (for tests)

# ─── Data Generation ───────────────────────────────────────────────────────────
faker==20.1.0

# ─── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0

# ─── Dev Tools ─────────────────────────────────────────────────────────────────
black==23.11.0
isort==5.12.0
flake8==6.1.0
